requirements.txt
streamlit
fbprophet
yfinance
plotly 